// generated/slskd-api/models/search-response.ts
export interface SearchResponse {
    /**
     * The username of the user who responded.
     * @type {string}
     * @memberof SearchResponse
     */
    username?: string;
    /**
     * The search token.
     * @type {number}
     * @memberof SearchResponse
     */
    token?: number;
    /**
     * The number of files in this response.
     * @type {number}
     * @memberof SearchResponse
     */
    file_count?: number;
    /**
     * The number of locked files in this response.
     * @type {number}
     * @memberof SearchResponse
     */
    locked_file_count?: number;
    /**
     * Whether the user has a free upload slot.
     * @type {boolean}
     * @memberof SearchResponse
     */
    has_free_upload_slot?: boolean;
    /**
     * The user's upload speed.
     * @type {number}
     * @memberof SearchResponse
     */
    upload_speed?: number;
    /**
     * The user's queue length.
     * @type {number}
     * @memberof SearchResponse
     */
    queue_length?: number;
    /**
     * The list of files in this response.
     * @type {Array<FileModel>}
     * @memberof SearchResponse
     */
    files?: Array<FileModel>;
    /**
     * The list of locked files in this response.
     * @type {Array<FileModel>}
     * @memberof SearchResponse
     */
    locked_files?: Array<FileModel>;
}